requirements.txt
# Django main
Django==3.2.*
django-cleanup==6.0.0
django-debug-toolbar==3.2.1
django-extensions==2.1.9
django-fsm==2.8.1
django-log-request-id==1.3.2
django-pgcrypto-fields==2.6.0
django-storages==1.11.1
django-webpack-loader==0.7.0
djangorestframework==3.12.*
django-sslserver==0.22

# Production Server
gunicorn==19.9.*
Pillow==8.3.2

# Custom Pixeldust Package
https://pixeldust-pypi.s3.ap-south-1.amazonaws.com/pg-pixeldust-common-0.0.1.tar.gz


# Database
psycopg2-binary==2.8.*
python-decouple==3.5

# Misc
requests==2.27.*
whitenoise==5.2.0
